+++ Sources/SwiftUIKit/Views/MenuList/MenuListIcon.swift
@@ -9,10 +9,14 @@
 import SwiftUI
 
 /**
- This view is used by `MenuListItem` as the leading icon.
+ This view is used by `MenuListItem` as the leading icon. It
+ presses the icon down to a certain size.
  */
 public struct MenuListIcon<Icon: View>: View {
     
+    /**
+     Create an icon with a custom image or view.
+     */
     public init(_ icon: Icon) {
         self.icon = icon
     }
@@ -21,7 +25,16 @@ public struct MenuListIcon<Icon: View>: View {
     
     public var body: some View {
         icon.frame(width: 20)
-            .padding(.trailing, 10)
+    }
+}
+
+public extension MenuListIcon where Icon == Text {
+    
+    /**
+     Create an icon with an emoji character.
+     */
+    init(_ emoji: String) {
+        self.init(Text(emoji))
     }
 }